Creating models.py:
I have made a menuItem class which has an id, name,
price, description and type attributes.
This will create the menuItem table in the database.db file.
This table uses id as the primary key.
It uses enum type to help differentiate between food and drinks.

I have made a Order class which has an id, order_total
and status attributes.
This will create the Order table in the database.db file.
It uses the id attribute as primary key.
It uses enum status to help differentiate between complete and
incomplete orders.