Creating branch featureOrderQuantityPrice -
I have updated models.py to allow Orders to hold duplicate menuitems.
I did this by creating a OrderMenuItem class that has a quantity field
and associates menuitems and orders together.

I then had to change customer.py add_to_cart method. This is where I
used the new OrderMenuItem class and checked if it already existed or not
and incremented the quantity of each menuitem accordingly.

I then had to change the route in customer.py for view_all_orders. This
 is where i passed both orders and menu items. Then in the view-all-orders
 html file, I could then compare the menu_items.id from Orders to menu_items
 and display the menu item name and quantity accordingly.