Creating branch featureOrderConfirmation to show order confirmation instead of redirecting customer to the home page.
This will allow the customer to check the order that they placed.
I used the bootstrap table used in waiter-view-order to similarly
display the customers order, with menu items and their quantity and then
the order total in the center.

To do this i had to create another route which queries the orderid,
ordermenuitem and menuitems and then passes these queries to the order-confirmation
template. I also had to then redirect to correct method in the add to cart
method.