Updating the add_item.html to include the types of each menu_item in the database.
This is done in waiter.py by querying the type attribute of each
menu_item using the distinct function. This is then passed to the
add_item template,where it does a for loop iterating through the types.
I used the index of each type to access the string value, otherwise it
wouldn't display the types.