Creating branch featureTableAssignment to allow - waiters to be asigned to a range of table_numbers.
After some research and testing, there was no List option i could find
that would work. I believe this way should be easy to handle regardless.

Whereby, each waiter will have a table_number_start and table_number_end.
This will be the range of tables which they will be in charge of.
When handling the waiters, a query can be made to see the last added waiters
table_number_end and their table_number_start can be that +1.

A simple range can be implemented to show only orders that the waiter is
assigned to in waiter.py